code/aleksandr-k3s-highload2/requirements.txt
Flask==2.2.5
requests==2.31.0
gunicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

